# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dilshod's API Gateway",
        "title": "API Gateway",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/goal/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new goal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Goal"
                ],
                "summary": "Create Goal",
                "parameters": [
                    {
                        "description": "Goal details",
                        "name": "goal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/goals.CreateGoalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Goal created successfully",
                        "schema": {
                            "$ref": "#/definitions/goals.GoalResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error while creating goal",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/goal/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a goal by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Goal"
                ],
                "summary": "Delete Goal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Goal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Goal deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/goals.GoalResponse"
                        }
                    },
                    "400": {
                        "description": "Missing or invalid ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error while deleting goal",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/goal/get/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve details of a goal by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Goal"
                ],
                "summary": "Get Goal by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Goal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Goal details",
                        "schema": {
                            "$ref": "#/definitions/goals.GetGoalResponse"
                        }
                    },
                    "400": {
                        "description": "Missing or invalid ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error while fetching goal",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/goal/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all goals based on the provided query parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Goal"
                ],
                "summary": "List Goals",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"user123\"",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"Save for vacation\"",
                        "description": "Name of the goal",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "example": 5000,
                        "description": "Target amount",
                        "name": "target_amount",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "example": 1500,
                        "description": "Current amount",
                        "name": "current_amount",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"2024-12-31\"",
                        "description": "Deadline (YYYY-MM-DD)",
                        "name": "deadline",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"In Progress\"",
                        "description": "Status of the goal",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of goals",
                        "schema": {
                            "$ref": "#/definitions/goals.ListGoalsResponse"
                        }
                    },
                    "500": {
                        "description": "Error while fetching goals",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/goal/update": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update details of a goal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Goal"
                ],
                "summary": "Update Goal",
                "parameters": [
                    {
                        "description": "Updated goal details",
                        "name": "goal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/goals.UpdateGoalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Goal updated successfully",
                        "schema": {
                            "$ref": "#/definitions/goals.GoalResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error while updating goal",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/notifications/delete": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a notification by user_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notification"
                ],
                "summary": "Delete Notification",
                "responses": {
                    "200": {
                        "description": "Notification deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/notifications.NotificationsResponse"
                        }
                    },
                    "404": {
                        "description": "Notification not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error while deleting notification",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/notifications/get": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a notification by user_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notification"
                ],
                "summary": "Get Notification",
                "responses": {
                    "200": {
                        "description": "Notification retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/notifications.GetNotificationByidResponse"
                        }
                    },
                    "404": {
                        "description": "Notification not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error while retrieving notification",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/notifications/getlist": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of all notifications",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notification"
                ],
                "summary": "List Notifications",
                "responses": {
                    "200": {
                        "description": "List of notifications retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/notifications.ListNotificationResponse"
                        }
                    },
                    "500": {
                        "description": "Error while retrieving notifications",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "goals.CreateGoalRequest": {
            "type": "object",
            "properties": {
                "current_amount": {
                    "type": "number"
                },
                "deadline": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "target_amount": {
                    "type": "number"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "goals.GetGoalResponse": {
            "type": "object",
            "properties": {
                "current_amount": {
                    "type": "number"
                },
                "deadline": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "target_amount": {
                    "type": "number"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "goals.GoalResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "goals.ListGoalsResponse": {
            "type": "object",
            "properties": {
                "goals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/goals.GetGoalResponse"
                    }
                }
            }
        },
        "goals.UpdateGoalRequest": {
            "type": "object",
            "properties": {
                "current_amount": {
                    "type": "number"
                },
                "deadline": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "target_amount": {
                    "type": "number"
                }
            }
        },
        "notifications.GetNotificationByidResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "notifications.ListNotificationResponse": {
            "type": "object",
            "properties": {
                "notifications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/notifications.GetNotificationByidResponse"
                    }
                }
            }
        },
        "notifications.NotificationsResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}